Added getOpts functionality.
Script no longer asks user if they want to re-run within a Screen session. Use the '-s' option when calling the script instead.

Added getOpts functionality. Flags for 'help', 'list' and 'screen' (-h, -l and -s).

Added date and time to the Screen log file.

Formatted help printout.

+++ wget-list-download.sh
@@ -2,37 +2,44 @@
 #
 # Download and rename items using wget from a list within a file. 
 #
-# See HelpMenu function for usage and list formatting. 
-#
-# Updated 2019-02-25
+# Run with '-h' to view help printout.  
 
-function HelpMenu(){
-	printf "\nHelp:\n"
-	printf "  Script usage: ./wget-list-download.sh listFilename\n"
-	printf "  File list format: filename.ext|download-url\n\n"
-	printf "  To call this help menu again, use '--help'\n\n"
+screenFlag='false'
+downloadList=''
+parameters='' 
+
+function CheckOptions(){
+	while getopts 'shl:' flag; do
+	  case "${flag}" in
+		s) screenFlag='true' ;;
+		h) PrintHelp && exit 0 ;;
+		l) downloadList="${OPTARG}" ;;
+		*) PrintHelp && exit 0 # If there's an argument we don't recognize, exit. 
+		exit 1 ;;
+	  esac
+	done	
+	
+	
+	if [ $# -eq 0 ]; then # If no arguments are passed, print help and exit. 
+		printf "\nNo argument specified.\n"
+		PrintHelp
+		exit 1
+	elif [ -z $downloadList ]; then # If there is no list file specified. 
+		printf "\nNo list file specified.\n"
+		PrintHelp
+		exit 1
+	fi
 }
 
-function CheckParam(){
-	if [ "$1" == "--help" ]; then # Help printout.
-		HelpMenu
-		exit 0
-	elif [ "$#" -eq 0 ]; then # If there isn't an input parameter. 
-		printf "No input file specified!\n"
-		HelpMenu
-		exit 1
-	elif [ "$#" -gt 1 ]; then # If there are too many input parameters.
-		printf "Too many parameters specified!\n"
-		HelpMenu
-		exit 1
-	elif [ "$#" -eq 1 ] && [ ! -f "$1" ]; then # If the input parameter is not a regular file. 
-		printf "Input parameter is not a regular file!\n"
-		HelpMenu
-		exit 1
-	elif [ "$#" -eq 1 ] && [ -f "$1" ]; then # If the input parameter is a single, regular file.
-		printf "\n"
-		downloadList="$1"
-	fi
+function PrintHelp(){
+	printf "\nHelp:\n"
+	printf "  Usage: ./wget-list-download.sh -l listFile.ext\n\n"
+	printf "  File list format: filename.ext|download-url\n\n"
+	printf "  Options:\n"
+	printf "    -h                Print help menu.\n"
+	printf "    -l [listFile]     REQUIRED. Specify the location of the list file.\n"
+	printf "    -s                Run the downloads within a Screen session.\n"
+	printf "\n"
 }
 
 function GetScreenInstallStatus(){
@@ -45,43 +52,51 @@ function GetScreenInstallStatus(){
 }
 
 # Check if this script is already in a Screen session.
-function AskForScreen(){
-	if GetScreenInstallStatus; then # If the screen package is installed. 
-		if [ ! -n "$STY" ]; then # If this script isn't in a Screen session.
-			while true; do
-				read -p "Do you want to re-open this script in a Screen session? (y/n) " -e screenSession # Ask for user input.
-				case $screenSession in
-					[Yy]* ) ReopenInScreen;; # If we do want to run in a Screen Session.
-					[Nn]* ) printf "\n" && break;; # If we don't, just continue.
-					* ) printf "\nInvalid input\n";;
-				esac
-			done
-		else # If this script is in a Screen session.
-			printf "Screen usage:\n"
-			printf "  To detach from a Screen session: CTRL A, then CTRL D\n"
-			printf "  To re-attach to this Screen session: 'screen -r wget-list-download'\n\n"
+function CheckScreen(){
+	if $screenFlag; then # If the Screen flag is true. 
+		if GetScreenInstallStatus; then # If the screen package is installed. 
+			if [ ! -n "$STY" ]; then # If this script isn't in a Screen session.
+				ReopenInScreen
+				exit 0
+			else # If this script is in a Screen session.
+				printf "Screen session name: $STY\n\n"
+				printf "Screen usage:\n"
+				printf "  To detach from a Screen session: CTRL A, then CTRL D\n"
+				printf "  To re-attach to this Screen session: 'screen -r wget-list-download'\n\n"
+			fi
+		else 
+			printf "Screen is not installed."
+			return # If screen isn't installed, just return and run without screen functions. 
 		fi
-	else 
-		return # If screen isn't installed, just return and run without screen functions. 
 	fi
 }
 
 function ReopenInScreen(){
-	exec screen -L -Logfile "$downloadList-wget-list-download.log" -S wget-list-download /bin/bash "$0" "$downloadList"; # Start a new instance of this script with same parameters within a Screen session... 
-	exit 0 # ... and exit this instance. 
+	dateTimeString=$(date +'%Y%m%d-%H%M%S')
+	printf "Re-opening in a Screen session. \nLog file: $dateTimeString-$downloadList-wget-list-download.log"
+	exec screen -L -Logfile "$dateTimeString-$downloadList-wget-list-download.log" -S wget-list-download /bin/bash "$0" "$parameters"; # Start a new instance of this script with same parameters within a Screen session... 
 }
 
 function DownloadFiles(){
+	printf "\nDownloading files from list: $downloadList\n\n"
 	if [ ! -z "$downloadList" ]; then # Check that the $downloadList variable isn't empty. 
-		while IFS=' | ' read -r fileName url; do
-			printf "Filename: $fileName\n"
-			wget -q --show-progress -O "$fileName" "$url"
-			printf "\n"
-		done < "$downloadList"
+		if [ -f "$downloadList" ]; then # Check if the list file parameter is a real file. 
+			while IFS='|' read -r fileName url; do
+				printf "Filename: $fileName\n"
+				wget -q --show-progress -O "$fileName" "$url"
+				printf "\n"
+			done < "$downloadList"
+		else 
+			printf "\nThe list file is not readable.\n"
+		fi
+	else 
+		printf "\nNo list file supplied. Exiting.\n"
+		exit 1
 	fi
 }
 
-CheckParam $@ # Check the parameters supplied, with all script paramaters as an input.
-AskForScreen # Check if the user has the screen package installed. If so, ask if they want to run this script is a Screen session.
-DownloadFiles # Download the files. 
+parameters="$@" # Store the the parameters that were passed to this script, as getopts within the CheckOptions function clears $@.
+CheckOptions $@
+CheckScreen
+DownloadFiles
 exit 0
